SUGGESTION: KAFKA-8147: Add changelog topic configuration to KTable suppress
Suggestion for https://github.com/apache/kafka/pull/8029
